+++ timeline.php
@@ -16,9 +16,6 @@ $con = new mysqli($dbhost, $dbuser, $dbpass, $dbname);
 if ($con->connect_error) {
    die("Connection failed: " . $con->connect_error); }
 
-
-
-
 $friends=friends::get_friends($con, $currentUSerID);
 
 array_push($friends, $currentUSerID) ;
@@ -26,30 +23,74 @@ array_push($friends, $currentUSerID) ;
 
 if(isset($_POST['txtInput']))
 {
-        $InputTxt=$_POST['txtInput']; 
-
-         header("Location:search.php?txt=".$InputTxt."");
-
-       
+    $InputTxt=$_POST['txtInput'];
+    header("Location:search.php?txt=".$InputTxt."");       
 }
 
 
-if(isset($_GET['likepostid'])) {
-    $like_postId= $_GET['likepostid'] ;
+if(isset($_POST['likepostid'])) {
+    $like_postId= $_POST['likepostid'] ;
     posts::like_post($con,$currentUSerID,$like_postId);
 }
+
+if(isset($_POST['comment']))
+{
+
+    echo posts::createComment($con,$_POST['comment_body'],$_POST['comment'], $currentUSerID);
+}
+
+if(isset($_POST['comments']))
+{
+    $post_id= $_POST['comments'] ;
+     $Comments = posts::display_comments($con,$post_id);
+    echo "<div onload = \"showModal()\"";
+           echo'<div class="modal fade" id="Modal" role="dialog" tabindex="-1" style="padding-top:100px; "overflow:scroll; height:400px;"" >
+        <div class="modal-dialog" role="document">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">×</span></button>
+                    <h4 class="modal-title">Comments</h4></div>
+                <div class="modal-body" style="max-height: 400px; overflow-y: auto">
+';
+                    
+                    if($Comments != false)
+                    {
+                       foreach ($Comments as $Comment) 
+                        {
+                            if($Comment['ID']==$currentUSerID)
+                            {
+                                echo "<blockquote><div> <a href=\"profile.php?id=".$Comment['ID']."\"> <h4 style=\" color:#337ab7; width:90%; text-transform: capitalize; font-size: 88%;\"> ~ ".$Comment['username']."</h4> </a> <button type=\"button\" class=\"close deletepost\" data-dismiss=\"modal\" aria-label=\"Close\" delete-commentid=".$Comment['comment_id'] ." postID = ".$post_id."><span aria-hidden=\"true\" style=\" font-size : 60%; color:red;\">Remove</span></button></button> </div> <p>".($Comment['body'])."</p><footer>Posted on ".
+                                $Comment['posted_at']."</blockquote><br> ";
+                            }
+
+                            else
+                            {
+                                echo "<blockquote><div> <a href=\"profile.php?id=".$Comment['ID']."\"> <h4 style=\" color:#337ab7; width:90%; text-transform: capitalize; font-size: 88%;\"> ~ ".$Comment['username']."</h4> </a> </div> <p>".($Comment['body'])."</p><footer>Posted on ".
+                                $Comment['posted_at']."</blockquote><br> ";
+                            }
+                            }
+                    }
+
+                echo' </div>
+                <div class="modal-footer">
+                    <button class="btn btn-default" type="button" data-dismiss="modal">Close</button>
+                </div>
+            </div>
+        </div>
+    </div>';
+}
     
-if(isset($_GET['likers'])) 
+if(isset($_POST['likers'])) 
     {
-        echo "<div onload = \"showlikersModal()\"";
-           echo'<div class="modal fade" id="likers" role="dialog" tabindex="-1" style="padding-top:100px;">
+        echo "<div onload = \"showModal()\"";
+           echo'<div class="modal fade" id="Modal" role="dialog" tabindex="-1" style="padding-top:100px;">
         <div class="modal-dialog" role="document">
             <div class="modal-content">
                 <div class="modal-header">
                     <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">×</span></button>
                     <h4 class="modal-title">Likers</h4></div>
                 <div class="modal-body" style="max-height: 400px; overflow-y: auto">';
-                    $like_postId= $_GET['likers'] ;
+                    $like_postId= $_POST['likers'] ;
                     $likers = posts::likers_IDs($con,$like_postId);
                     if($likers != false)
                     {
@@ -76,17 +117,22 @@ if(isset($_GET['likers']))
 
     }
 
-    
-?> 
+if(isset( $_GET['deletepostid']))
+{ 
+    $delete_postId= $_GET['deletepostid'] ;
+    $message = posts::delete_post($con,$currentUSerID,$delete_postId);
+}
 
-<?php  
-        $delete_postId= $_GET['deletepostid'] ;
-        $message = posts::delete_post($con,$currentUSerID,$delete_postId);
-
+if(isset($_GET['deletecommentid']))
+{
+    $delete_commentId= $_GET['deletecommentid'] ;
+    $commentPostID = $_GET['postID'] ;
+    $message = posts::deleteComment ($con, $delete_commentId ,$commentPostID);
+}
 ?>
 
 <!DOCTYPE html>
-<html>
+<html > 
 
 <head>
     <meta charset="utf-8">
@@ -101,8 +147,7 @@ if(isset($_GET['likers']))
     <link rel="stylesheet" href="assets/css/styles.css">
     <link rel="stylesheet" href="assets/css/untitled.css">
 </head>
-
-<body onload="showlikersModal()">
+<body onload="showModal()">
     <header class="hidden-sm hidden-md hidden-lg">
         <div class="searchbox">
             <form validate method="post">
@@ -156,7 +201,7 @@ if(isset($_GET['likers']))
             </div>
         </nav>
     </div>
-    <div class="container" style="margin-bottom: 60px;">
+    <div  class="col-md-8" class="container" style="margin-bottom: 60px; margin-left: 10%;">
         <h1>Timeline </h1>
         <div class="timelineposts" >
 
@@ -164,22 +209,7 @@ if(isset($_GET['likers']))
 
         </div>
     </div>
-    <div class="modal fade" role="dialog" tabindex="-1" style="padding-top:100px;">
-        <div class="modal-dialog" role="document">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">×</span></button>
-                    <h4 class="modal-title">Comments</h4></div>
-                <div class="modal-body" style="max-height: 400px; overflow-y: auto">
-                    <p>The content of your modal.</p>
-                </div>
-                <div class="modal-footer">
-                    <button class="btn btn-default" type="button" data-dismiss="modal">Close</button>
-                </div>
-            </div>
-        </div>
-    </div>
-
+    <!-- Comments -->    
     <div class="footer-dark navbar-fixed-bottom" style="position: fixed;">
         <footer>
             <div class="container">
@@ -199,13 +229,23 @@ if(isset($_GET['likers']))
                              
         });
 
-        function showlikersModal() {
-                            $('#likers').modal('show');} 
+         $('[delete-commentid]').click(function() {
+                 var buttonid = $(this).attr('delete-commentid');
+                 var postid = $(this).attr('postID');
+                 window.location.replace('timeline.php?deletecommentid='+ buttonid +'&&postID='+postid);
+                             
+        });
 
+        function showModal() 
+
+
+
+        {
+            $('#Modal').modal('show');
+        } 
+       
         
     </script>
-
-
 
 </body>
